+++ app/src/main/kotlin/org/btcmap/imagestats/ImageDiskCacheStats.kt
@@ -1,0 +1,8 @@
+package org.btcmap.imagestats
+
+/** Usage of Coil's on-disk image cache. */
+data class ImageDiskCacheStats(
+    val directory: String,
+    val usedBytes: Long,
+    val maxBytes: Long,
+)